Moved map to user folder
