+++ src/App.jsx
@@ -24,7 +24,7 @@ import InsuranceCompanies from './components/pages/insurance/InsuranceCompanies'
 import CompanyPage from './components/pages/insurance/CompanyPage'
 import InsuranceCardPage from './components/pages/insurance/InsuranceCardPage'
 import NewPassword from './components/pages/login/NewPassword'
-import Map from './components/pages/map/Map'
+import Map from './components/pages/user/map/Map'
 import AddCard from './components/pages/insurance/AddCard'
 import CategoryPage from './components/pages/categoryPage/CategoryPage'
 import ErrorPage from './components/common/ErrorPage'
